first project displays demo

+++ StartAppBackend/src/main/java/com/skyletto/startappbackend/utils/Utils.java
@@ -1,7 +1,7 @@
 package com.skyletto.startappbackend.utils;
 
 public class Utils {
-    static public double diff(int zoom){
+    static public double diff(float zoom){
         return 0.06*((20-zoom)*(20-zoom));
     }
 }
