+++ StartAppFrontend/app/src/main/java/com/skyletto/startappfrontend/data/requests/LatLngRequest.kt
@@ -1,3 +1,3 @@
 package com.skyletto.startappfrontend.data.requests
 
-data class LatLngRequest(var lat: Double, var lng: Double, var zoom:Int = 16)
+data class LatLngRequest(var lat: Double, var lng: Double, var zoom:Float = 16f)
